Added three new languages: French, Italian and Portuguese. These are the three other languages supported by Alta Vista's Babelfish.
Also added other messages to the German module.

One minor edit in the Spanish module.

Default: Fixed a couple of typos and added some material at the top that
shows the source material I handed to Babelfish to get the translations.

I did my best to reality-check the output before copying it in, but I don't
speak any of these other than Spanish.